refactor: DirectoryNode and FileNode types
