+++ src/components/FileRow.tsx
@@ -1,7 +1,7 @@
+import { FileNode } from "@/types";
 import { PencilSquareIcon } from "@heroicons/react/24/outline";
 import Link from "next/link";
 import { FC } from "react";
-import { FileNode } from "./FileTree";
 
 type FileRowProps = {
   fileNode: FileNode;
